# Conflicts:
#	logs/springboot-base-frame-error.log
#	logs/springboot-base-frame-warn.log
#	src/main/java/com/phoenix/free/dto/AddPartnerPost.java
#	src/main/java/com/phoenix/free/entity/Message.java

+++ src/main/java/com/phoenix/free/service/RelationshipService.java
@@ -5,7 +5,7 @@ import com.phoenix.free.entity.Relationship;
 import java.util.List;
 
 public interface RelationshipService {
-    int addNewRelationship(Long userId1, Long userId2);
+    Long addNewRelationship(Long userId1, Long userId2);
     void deleteRelationship(Long userId1, Long userId2);
 
     List<Relationship> getUserRelationship(Long userId);
